fec_mxc: change data argument of fecmxc_smi_write() to uint8_t
SMI only deals with 8-bit registers, the upper 8 bit are ignored.
Reflect this in the type signature and remove the meaningless digits
from the EM310 fecmxc_mii_postcall().

+++ drivers/net/fec_mxc.c
@@ -83,7 +83,7 @@ static void swap_packet(uint32_t *packet, int length)
 
 #ifndef CONFIG_PHYLIB
 /* SMI-interface related functions */
-int fecmxc_smi_write(struct eth_device *dev, uint8_t regAddr, uint16_t data)
+int fecmxc_smi_write(struct eth_device *dev, uint8_t regAddr, uint8_t data)
 {
 	struct fec_priv *fec = (struct fec_priv *)dev->priv;
 	struct ethernet_regs *eth = fec->bus->priv;
